feat: add destination_hotkey parameter to transfer_stake
Closes #1377

This PR allows users to specify a destination hotkey when transferring stake
to a different coldkey. Previously, transfer_stake only allowed changing the
coldkey while keeping the same hotkey, which caused issues because the new
coldkey wouldn't own the original hotkey.

Changes:
- Add destination_hotkey parameter to transfer_stake dispatch (call_index 86)
- Update do_transfer_stake to use separate origin_hotkey and destination_hotkey
- Update StakeTransferred event to include both hotkeys
- Update precompiles (transferStake now takes 6 params instead of 5)
- Update chain-extensions for new signature
- Update contract interface (ink! contracts)
- Update solidity interface
- Add test for transferring to different hotkey
- Update all existing tests to use new parameter
- Update documentation

This is a breaking change for existing callers of transfer_stake.